agent/pyproject.toml
[tool.poetry]
name = "agent"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
mem0ai = "^0.1.2"
openai = "^1.42.0"
twilio = "^9.2.3"
flask = "^3.0.3"
python-dotenv = "^1.0.1"
pyngrok = "^7.2.0"
pydantic = "^2.8.2"
flask-pydantic = "^0.12.0"
devtools = "^0.12.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
